Separate environment variable for project path
On gitlab, the namespace need not be a user. Sometimes it can be a
group, which can have several nesting levels. Also, sometimes a gitlab
user might not run this script against a project under her name, so I
think it's better to use a separate variable for the project's path.