Add small pause before sending auto merge request
I have consistently come across a 405 when using this script with merge_when_pipeline_succeeds. This seems to be a common issue even with the multiple fixes that have been provided. From what I have seen, the 405 is from accepting the MR before its ready. This pause - which may be able to be shortened - has allowed me to utilize the merge_when_pipeline_succeeds flag on my Gitlab instance.